Add TileBitmaskCore library and refactor projects
