+++ Bitmask.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.14.36127.28
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TileBitmaskGen", "TileBitmaskGen\TileBitmaskGen.csproj", "{C1EFB28E-002D-4381-9020-6964944672B9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TileBitmaskCore", "TileBitmaskCore\TileBitmaskCore.csproj", "{88645AFD-4C29-4996-8B49-317572FC7B95}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestProject", "TestProject\TestProject.csproj", "{54450E9C-52E2-4443-86A2-8383CA83268F}"
 EndProject
 Global
@@ -17,6 +19,10 @@ Global
 		{C1EFB28E-002D-4381-9020-6964944672B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C1EFB28E-002D-4381-9020-6964944672B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1EFB28E-002D-4381-9020-6964944672B9}.Release|Any CPU.Build.0 = Release|Any CPU
+                {88645AFD-4C29-4996-8B49-317572FC7B95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {88645AFD-4C29-4996-8B49-317572FC7B95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {88645AFD-4C29-4996-8B49-317572FC7B95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {88645AFD-4C29-4996-8B49-317572FC7B95}.Release|Any CPU.Build.0 = Release|Any CPU
 		{54450E9C-52E2-4443-86A2-8383CA83268F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{54450E9C-52E2-4443-86A2-8383CA83268F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{54450E9C-52E2-4443-86A2-8383CA83268F}.Release|Any CPU.ActiveCfg = Release|Any CPU
