+++ TestProject/TestProject.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net8.0-windows</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TileBitmaskGen\TileBitmaskGen.csproj" />
+    <ProjectReference Include="..\TileBitmaskCore\TileBitmaskCore.csproj" />
   </ItemGroup>
 
   <ItemGroup>
