+++ TileBitmaskCore/TileBitmaskStringGenerator.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TileBitmaskGen
+namespace TileBitmaskCore
 {
     internal class TileBitmaskStringGenerator
     {
