+++ TileBitmaskCore/TileRule.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TileBitmaskGen
+namespace TileBitmaskCore
 {
     public class TileRule
     {
